# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kayak rental platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /catalog:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: List all active products
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: capacity
          schema:
            type: string
        - in: query
          name: construction_type
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /catalog/{slug}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: Get product by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /chat/session:
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/session/{sessionId}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get messages for a session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
        - in: query
          name: since
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message in a session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
  /forms/cheaper-price:
    post:
      operationId: submitCheaperPrice
      tags: [forms]
      summary: Submit "found it cheaper" form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheaperPriceFormInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /forms/feedback:
    post:
      operationId: submitFeedback
      tags: [forms]
      summary: Submit feedback form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackFormInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        sku:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        shortDescription:
          type: ["string", "null"]
        capacity:
          type: ["integer", "null"]
        constructionType:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        depositAmount:
          type: ["number", "null"]
        featured:
          type: boolean
        badge:
          type: ["string", "null"]
        mainImage:
          type: ["string", "null"]
        tariffs:
          type: array
          items:
            $ref: "#/components/schemas/Tariff"
      required:
        - id
        - name
        - slug
        - categoryId
        - categoryName
        - featured
    ProductDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        sku:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        shortDescription:
          type: ["string", "null"]
        fullDescription:
          type: ["string", "null"]
        capacity:
          type: ["integer", "null"]
        constructionType:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        kit:
          type: ["string", "null"]
        depositAmount:
          type: ["number", "null"]
        featured:
          type: boolean
        badge:
          type: ["string", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        h1:
          type: ["string", "null"]
        ogTitle:
          type: ["string", "null"]
        ogDescription:
          type: ["string", "null"]
        canonicalUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        tariffs:
          type: array
          items:
            $ref: "#/components/schemas/Tariff"
        specifications:
          type: object
          additionalProperties: true
      required:
        - id
        - name
        - slug
        - categoryId
        - categoryName
        - featured
        - images
        - tariffs
    ProductImage:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        alt:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - id
        - url
        - sortOrder
    Tariff:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [weekend, week, weekday, may_holidays]
        label:
          type: string
        pricePerDay:
          type: number
        minDays:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
      required:
        - id
        - type
        - label
        - pricePerDay
...
    ChatSession:
      type: object
      properties:
        id:
          type: integer
        sessionToken:
          type: string
        status:
          type: string
        customerId:
          type: ["integer", "null"]
        orderId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - sessionToken
        - status
        - createdAt
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        sender:
          type: string
          enum: [customer, manager, bot]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - sender
        - content
        - createdAt
...
    CreateChatSessionRequest:
      type: object
      properties:
        customerName:
          type: ["string", "null"]
        orderId:
          type: ["integer", "null"]
      required: []
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...
    CheaperPriceFormInput:
      type: object
      properties:
        productId:
          type: ["integer", "null"]
        name:
          type: string
        contact:
          type: string
        competitorUrl:
          type: ["string", "null"]
        competitorPrice:
          type: ["number", "null"]
        comment:
          type: ["string", "null"]
      required:
        - name
        - contact
    FeedbackFormInput:
      type: object
      properties:
        name:
          type: string
        contact:
          type: string
        message:
          type: string
        pageUrl:
          type: ["string", "null"]
      required:
        - name
        - contact
        - message